+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+<<<<<<< HEAD
+=======
+## 0.2.0 - (2022-04-11)
+
+* Return sorted integer scenarios ids
+* Unrecognized endpoints treated as warnings, not errors
+
+>>>>>>> release
 ## 0.1.0 - (2022-04-06)
 
 * First beta release